# docs/swagger.yaml
basePath: /
definitions:
  db.Subscription:
    properties:
      ends_At:
        type: string
      id:
        type: integer
      price:
        type: integer
      service:
        type: string
      started_At:
        type: string
      userID:
        type: string
    type: object
  usecase.SubscriptionReq:
    properties:
      ends_at:
        type: string
      price:
        type: integer
      service:
        type: string
      started_at:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST-сервис для работы с подписками
  title: Subscription API
  version: "1.0"
paths:
  /subscriptions:
    get:
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Subscription'
            type: array
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/usecase.SubscriptionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.Subscription'
        "400":
          description: невалидный JSON / поля
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Создание подписки
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: удалено
          schema:
            type: string
        "404":
          description: не найдено
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Удаление подписки
      tags:
      - subscriptions
    get:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Subscription'
        "404":
          description: не найдено
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Получить подписку по ID
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: patch
        required: true
        schema:
          type: object
      responses:
        "204":
          description: обновлено
          schema:
            type: string
        "400":
          description: невалидный JSON
          schema:
            type: string
        "404":
          description: не найдено
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Частичное обновление подписки
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/usecase.SubscriptionReq'
      responses:
        "204":
          description: обновлено
          schema:
            type: string
        "400":
          description: невалидный JSON
          schema:
            type: string
        "404":
          description: не найдено
          schema:
            type: string
        "500":
          description: внутренняя ошибка
          schema:
            type: string
      summary: Полное обновление подписки
      tags:
      - subscriptions
swagger: "2.0"